+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+
+## Unreleased
+### Fixed
+- Integration tests
+
 ## [1.10.2-rc.0] - 2019-09-16
 ### Added 
 - Builder for `RemoveAssetOp`