+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## Unreleased 
+### Fixed
+- tests for atomic swaps
+
 ## [1.10.2-rc.1] - 2019-09-19
 ### Added
 - Builder for `OpenSwapOp`